Feature: link gui

+++ src/main/java/fr/ul/acl/escape/monde/Monstre.java
@@ -1,7 +1,7 @@
 package fr.ul.acl.escape.monde;
 
 public abstract class Monstre extends Personnage {
-    public Monstre(float x, float y, int hauteur, int largeur, float vitesse) {
+    public Monstre(double x, double y, double hauteur, double largeur, double vitesse) {
         super(x, y, hauteur, largeur, vitesse);
     }
 }
