+++ src/main/java/fr/ul/acl/escape/monde/Terrain.java
@@ -1,7 +1,7 @@
 package fr.ul.acl.escape.monde;
 
 public abstract class Terrain extends ElementMonde {
-    public Terrain(float x, float y, int hauteur, int largeur) {
+    public Terrain(double x, double y, double hauteur, double largeur) {
         super(x, y, hauteur, largeur);
     }
